Set up conformance tests (#6)
* Enhanced testing setup in language package

The test setup in the language package has been improved. A context parameter was added to the BeforeEach function, and a call to GetLanguageTests was included. This ensures that all necessary tests are fetched before each test run. Additionally, a comment was added to indicate the need for reducing global state usage in the future.

* Updated language host visibility and GRPC client creation

The visibility of the function creating a new language host has been changed from private to public. This allows other packages to create a new language host. Additionally, the method for creating a GRPC client connection has been simplified by using `grpc.NewClient` instead of `grpc.Dial`. A comment was also added in the test file regarding dynamic spec generation.

* Ignore user VSCode settings

Ignore test output

* Don't shoot myself in the foot and just do it how pulumi does it

* Wrap up the test setup

* Point to the vendored nodejs SDK as a starting point

* Stub out pack so we can at least get to the tests